refactor: mover classification_legend() de banks.py para app.py
