+++ app.py
@@ -10,6 +10,17 @@ st.set_page_config(page_title='easy-financ-export', layout='centered')
 Navbar()
 
 
+def classification_legend():
+    st.markdown(
+        '<small>'
+        '<span style="color:gray">■ Regras</span>&nbsp;&nbsp;'
+        '<span style="color:#1565C0">■ Histórico</span>&nbsp;&nbsp;'
+        '<span style="color:#E65100">■ Modelo</span>'
+        '</small>',
+        unsafe_allow_html=True,
+    )
+
+
 # --- XP Investimentos ---
 
 try:
@@ -24,7 +35,7 @@ try:
 
     if st.toggle("*Classificar transações ?*", value=False, key='xp_classifier'):
         xp_class = banks.classify_xp(xp)
-        banks.classification_legend()
+        classification_legend()
         st.dataframe(banks.style_classified(banks.display_xp(xp_class)))
     else:
         st.dataframe(banks.display_xp(xp))
@@ -72,7 +83,7 @@ try:
         nubank = classifier.classify_complete(nubank, numeric_col='Valor', cat_col='Estabelecimento')
         st.metric("Valor Parcial", round(nubank['Valor'].astype('float64').sum(), 2))
         nubank['Valor'] = nubank['Valor'].astype('str').str.replace('.', ',')
-        banks.classification_legend()
+        classification_legend()
         st.dataframe(banks.style_classified(nubank))
     else:
         st.metric("Valor Parcial", round(nubank['Valor'].astype('float64').sum(), 2))
